webapp/requirements.txt
python-dotenv>=1.0,<2.0
tqdm>=4.65,<5.0
